requirements.txt
django-jazzmin~=3.0.1
Django~=5.1.6
djangorestframework~=3.15.2
django-cleanup~=9.0.0
django-filter~=25.1
drf-yasg~=1.21.9
slugify~=0.0.1